cli/pyproject.toml
[project]
name = "dotfiles-cli"
version = "0.1.0"
description = "Personal dotfiles management CLI"
requires-python = ">=3.9"
dependencies = [
    "cyclopts>=3.0",
    "PyYAML>=6",
    "rich>=13",
]

[project.scripts]
dotfiles = "dotfiles_cli.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dotfiles_cli"]
